add support for addressing barcode sequences in output template
Molecular barcodes that have no been error corrected are now written to OX/BZ instead of RX/QX

Added support for s,c,m in the first literal of a token to address the decoded sample, cellular and molecular sequence. the decoded sequence is addressed as one, concatenated, sequence and not as a segmented sequence. For now the quality scores are unchanged from the observed sequence. This needs to be changed into something that makes more sense...

for instance `s::` will be the entire sample barcode sequence.

Fixed a bug in schema validation where template element was not validated.